requirements.txt
openai
pandas
python-dotenv
pydantic
faker
pyarrow
